[CDN-457] Added Codefresh Build
Adds a base Dockerfile which is used by codefresh to build Janus,
janus-plugin-forwarder, and create the artifacts bundle to be uploaded
to the S3 bucket.

NOTE: Unlike Jenkins, Codefresh does not provide an increasing build
number. Instead, it provides each build with a unique identifier. This
is used in place of the build number to tag build artifacts.